+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-### 0.2.0
-The biggest update so far!
+### 0.3.0
+- Added light theme (eww)
+- New `theme` command! Allows you to change the terminal theme.
+  - Usage: `theme NAME` (example: `theme light`)
+- Potentially fixed `changelog` command.
+- Updated MOTDs.
+
+### 0.2.0 - The biggest update so far!
 - `help` command can now show detailed info about most of the commands.
   - Usage: `help NAME` (example: `help project`)
 - New `project` command! Shows detailed info about my featured projects.
@@ -14,4 +20,3 @@ The biggest update so far!
 
 ### 0.1.0
 Initial release.
-- Changelog is WIP